requirements.txt
nltk~=3.4.5
tensorflow-gpu==2.0.1
deepsegment~=2.3.0
pyprind~=2.11.2
attrs~=19.1.0
torch>=1.2.0
numpy~=1.17.2
pandas~=0.25.1
spacy>=2.1.0,<2.2
https://github.com/phueb/Ludwig/archive/v2.1.3.tar.gz